refactor: rollback template implementation outside the jpa package
