+++ src/main/java/com/joutvhu/dynamic/jpa/handlebars/helpers/CustomDirectiveHelpers.java
@@ -1,8 +1,10 @@
-package com.joutvhu.dynamic.handlebars.helpers;
+package com.joutvhu.dynamic.jpa.handlebars.helpers;
 
 import com.github.jknack.handlebars.Handlebars;
 import com.github.jknack.handlebars.Helper;
 import com.github.jknack.handlebars.Options;
+import com.joutvhu.dynamic.commons.util.TrimProcessor;
+import com.joutvhu.dynamic.commons.util.TrimSymbol;
 import lombok.NoArgsConstructor;
 import lombok.SneakyThrows;
 
@@ -11,23 +13,20 @@ import java.io.IOException;
 import static org.apache.commons.lang3.Validate.notNull;
 
 @NoArgsConstructor
-public enum CustomStringHelpers implements Helper<Object> {
+public enum CustomDirectiveHelpers implements Helper<Object> {
 
-    startsWith {
+    where {
+        private final TrimProcessor processor = new TrimProcessor();
+        private final TrimSymbol symbols = new TrimSymbol(
+                "where", TrimSymbol.getOverrides(true, "and", "or"),
+                null, TrimSymbol.getOverrides(false, "and", "or")
+        );
+
         @Override
         @SneakyThrows
         protected CharSequence safeApply(final Object value, final Options options) {
-            boolean result = value.toString().startsWith(options.param(0));
-            return result ? options.fn() : options.inverse();
-        }
-    },
-
-    endsWith {
-        @Override
-        @SneakyThrows
-        protected CharSequence safeApply(final Object value, final Options options) {
-            boolean result = value.toString().endsWith(options.param(0));
-            return result ? options.fn() : options.inverse();
+            CharSequence content = options.fn();
+            return processor.process(symbols, content.toString());
         }
     };
 
@@ -66,8 +65,8 @@ public enum CustomStringHelpers implements Helper<Object> {
      */
     public static void register(final Handlebars handlebars) {
         notNull(handlebars, "A handlebars object is required.");
-        CustomStringHelpers[] helpers = values();
-        for (CustomStringHelpers helper : helpers) {
+        CustomDirectiveHelpers[] helpers = values();
+        for (CustomDirectiveHelpers helper : helpers) {
             helper.registerHelper(handlebars);
         }
     }
